refactor RootInvokeEvent
