+++ sharding-jdbc/src/main/java/io/shardingsphere/core/jdbc/core/connection/ShardingConnection.java
@@ -18,7 +18,7 @@
 package io.shardingsphere.core.jdbc.core.connection;
 
 import io.shardingsphere.core.event.ShardingEventBusInstance;
-import io.shardingsphere.core.event.root.RootInvokeEvent;
+import io.shardingsphere.core.event.root.RootInvokeStartEvent;
 import io.shardingsphere.core.jdbc.adapter.AbstractConnectionAdapter;
 import io.shardingsphere.core.jdbc.core.ShardingContext;
 import io.shardingsphere.core.jdbc.core.statement.ShardingPreparedStatement;
@@ -53,7 +53,7 @@ public final class ShardingConnection extends AbstractConnectionAdapter {
         super(shardingContext.getDatabaseType());
         this.dataSourceMap = dataSourceMap;
         this.shardingContext = shardingContext;
-        ShardingEventBusInstance.getInstance().post(new RootInvokeEvent());
+        ShardingEventBusInstance.getInstance().post(new RootInvokeStartEvent());
     }
     
     /**
